+++ CHANGELOG.md
@@ -17,6 +17,7 @@ See LICENSE.txt for license information.
 - Remove Home tab
 - Make Recommendations the primary tab
 - Improve readability of tab symbols in menu bar
+- Link directly to setting in Recommendations → Activation Lock
 
 ### Fixed